refactor: BaseService and BaseServiceTest use new status
