+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/enums/v2/ProjectStatus.java
@@ -17,21 +17,18 @@
 
 package org.apache.dolphinscheduler.api.enums.v2;
 
-import java.util.Locale;
-
 import org.springframework.context.i18n.LocaleContextHolder;
 
-public enum BaseStatus implements Status {
-    SUCCESS(0, "success", "成功"),
-    INTERNAL_SERVER_ERROR_ARGS(10000, "Internal Server Error: {0}", "服务端异常: {0}"),
-    REQUEST_PARAMS_NOT_VALID_ERROR(10001, "request parameter {0} is not valid", "请求参数[{0}]无效"),
-    CREATE_ACCESS_TOKEN_ERROR(70010, "create access token error", "创建访问token错误"),
-    ACCESS_TOKEN_NOT_EXIST(70015, "access token not exist, tokenId {0}", "访问token不存在, {0}"),;
+import java.util.Locale;
+
+public enum ProjectStatus implements Status {
+    PROJECT_NOT_FOUND(10018, "project {0} not found ", "项目[{0}]不存在"),
+    ;
     private final int code;
     private final String enMsg;
     private final String zhMsg;
 
-    BaseStatus(int code, String enMsg, String zhMsg) {
+    ProjectStatus(int code, String enMsg, String zhMsg) {
         this.code = code;
         this.enMsg = enMsg;
         this.zhMsg = zhMsg;
